rebalance

+++ src/main/java/com/teamwizardry/wizardry/api/WizardManager.java
@@ -17,67 +17,67 @@ public class WizardManager {
 	private WizardManager() {
 	}
 
-	public static void setMaxMana(int mana, @Nullable EntityLivingBase player) {
+	public static void setMaxMana(double mana, @Nullable EntityLivingBase player) {
 		IWizardryCapability cap = getCap(player);
 		if (cap == null) return;
 		cap.setMaxMana(mana, (EntityPlayer) player);
 		if (getMana(player) > mana) setMana(mana, player);
 	}
 
-	public static void setMana(int mana, @Nullable EntityLivingBase player) {
+	public static void setMana(double mana, @Nullable EntityLivingBase player) {
 		IWizardryCapability cap = getCap(player);
 		if (cap == null) return;
 		cap.setMana(Math.min(Math.max(0, mana), getMaxMana(player)), (EntityPlayer) player);
 	}
 
-	public static void addMana(int mana, @Nullable EntityLivingBase player) {
+	public static void addMana(double mana, @Nullable EntityLivingBase player) {
 		setMana(getMana(player) + mana, player);
 	}
 
-	public static void removeMana(int mana, @Nullable EntityLivingBase player) {
+	public static void removeMana(double mana, @Nullable EntityLivingBase player) {
 		setMana(getMana(player) - mana, player);
 	}
 
-	public static void setMaxBurnout(int burnout, @Nullable EntityLivingBase player) {
+	public static void setMaxBurnout(double burnout, @Nullable EntityLivingBase player) {
 		IWizardryCapability cap = getCap(player);
 		if (cap == null) return;
 		cap.setMaxBurnout(burnout, (EntityPlayer) player);
 		if (getBurnout(player) > burnout) setBurnout(burnout, player);
 	}
 
-	public static void setBurnout(int burnout, @Nullable EntityLivingBase player) {
+	public static void setBurnout(double burnout, @Nullable EntityLivingBase player) {
 		IWizardryCapability cap = getCap(player);
 		if (cap == null) return;
 		cap.setBurnout(Math.max(0, Math.min(burnout, getMaxBurnout(player))), (EntityPlayer) player);
 	}
 
-	public static void addBurnout(int burnout, @Nullable EntityLivingBase player) {
+	public static void addBurnout(double burnout, @Nullable EntityLivingBase player) {
 		setBurnout(getBurnout(player) + burnout, player);
 	}
 
-	public static void removeBurnout(int burnout, @Nullable EntityLivingBase player) {
+	public static void removeBurnout(double burnout, @Nullable EntityLivingBase player) {
 		setBurnout(getBurnout(player) - burnout, player);
 	}
 
-	public static int getMaxMana(@Nullable EntityLivingBase player) {
+	public static double getMaxMana(@Nullable EntityLivingBase player) {
 		IWizardryCapability cap = getCap(player);
 		if (cap == null) return 0;
 		return cap.getMaxMana();
 	}
 
-	public static int getMana(@Nullable EntityLivingBase player) {
+	public static double getMana(@Nullable EntityLivingBase player) {
 		IWizardryCapability cap = getCap(player);
 		if (cap == null) return 0;
 		return cap.getMana();
 	}
 
-	public static int getBurnout(@Nullable EntityLivingBase player) {
+	public static double getBurnout(@Nullable EntityLivingBase player) {
 		IWizardryCapability cap = getCap(player);
 		if (cap == null) return 0;
 		return cap.getBurnout();
 	}
 
-	public static int getMaxBurnout(@Nullable EntityLivingBase player) {
+	public static double getMaxBurnout(@Nullable EntityLivingBase player) {
 		IWizardryCapability cap = getCap(player);
 		if (cap == null) return 0;
 		return cap.getMaxBurnout();
