+++ src/main/java/com/teamwizardry/wizardry/api/capability/WizardryCapabilityStorage.java
@@ -20,10 +20,10 @@ public class WizardryCapabilityStorage implements IStorage<IWizardryCapability> 
 		NBTTagCompound nbt = new NBTTagCompound();
 		if (instance.getBloodType() != null)
 			nbt.setString(BLOOD_TYPE, instance.getBloodType().id);
-		nbt.setInteger(MAX_MANA, instance.getMaxMana());
-		nbt.setInteger(MAX_BURNOUT, instance.getMaxBurnout());
-		nbt.setInteger(MANA, instance.getMana());
-		nbt.setInteger(BURNOUT, instance.getBurnout());
+		nbt.setDouble(MAX_MANA, instance.getMaxMana());
+		nbt.setDouble(MAX_BURNOUT, instance.getMaxBurnout());
+		nbt.setDouble(MANA, instance.getMana());
+		nbt.setDouble(BURNOUT, instance.getBurnout());
 
 		return nbt;
 	}
@@ -31,10 +31,10 @@ public class WizardryCapabilityStorage implements IStorage<IWizardryCapability> 
 	@Override
 	public void readNBT(Capability<IWizardryCapability> capability, IWizardryCapability instance, EnumFacing side, NBTBase nbt) {
 		NBTTagCompound tag = (NBTTagCompound) nbt;
-		((DefaultWizardryCapability) instance).mana = tag.getInteger(MANA);
-		((DefaultWizardryCapability) instance).maxMana = tag.getInteger(MAX_MANA);
-		((DefaultWizardryCapability) instance).burnout = tag.getInteger(BURNOUT);
-		((DefaultWizardryCapability) instance).maxBurnout = tag.getInteger(MAX_BURNOUT);
+		((DefaultWizardryCapability) instance).mana = tag.getDouble(MANA);
+		((DefaultWizardryCapability) instance).maxMana = tag.getDouble(MAX_MANA);
+		((DefaultWizardryCapability) instance).burnout = tag.getDouble(BURNOUT);
+		((DefaultWizardryCapability) instance).maxBurnout = tag.getDouble(MAX_BURNOUT);
 		if (tag.hasKey(BLOOD_TYPE))
 			((DefaultWizardryCapability) instance).bloodType = EnumBloodType.getType(tag.getString(BLOOD_TYPE));
 	}
